requirements.txt
streamlit>=1.36.0
tensorflow-cpu==2.15.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python-headless>=4.9.0.80
pandas>=2.1.0
jsonschema>=4.19.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
